Move collection/product hooks
