+++ assets/js/base/context/hooks/collections/use-collection-header.js
@@ -3,11 +3,7 @@
  */
 import { COLLECTIONS_STORE_KEY as storeKey } from '@woocommerce/block-data';
 import { useSelect } from '@wordpress/data';
-
-/**
- * Internal dependencies
- */
-import { useShallowEqual } from './use-shallow-equal';
+import { useShallowEqual } from '@woocommerce/base-hooks';
 
 /**
  * This is a custom hook that is wired up to the `wc/store/collections` data
