+++ assets/js/base/context/hooks/collections/use-collection.js
@@ -4,12 +4,7 @@
 import { COLLECTIONS_STORE_KEY as storeKey } from '@woocommerce/block-data';
 import { useSelect } from '@wordpress/data';
 import { useRef } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
-import { useShallowEqual } from './use-shallow-equal';
-import { useThrowError } from './use-throw-error';
+import { useShallowEqual, useThrowError } from '@woocommerce/base-hooks';
 
 /**
  * This is a custom hook that is wired up to the `wc/store/collections` data
