+++ assets/js/base/context/hooks/index.js
@@ -1,3 +1,6 @@
 export * from './cart';
+export * from './collections';
 export * from './use-store-notices';
 export * from './use-store-events';
+export * from './use-query-state';
+export * from './use-store-products';
