+++ assets/js/base/context/hooks/use-query-state.js
@@ -4,14 +4,14 @@
 import { QUERY_STATE_STORE_KEY as storeKey } from '@woocommerce/block-data';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { useRef, useEffect, useCallback } from '@wordpress/element';
-import { useQueryStateContext } from '@woocommerce/base-context';
 import isShallowEqual from '@wordpress/is-shallow-equal';
+import { useShallowEqual, usePrevious } from '@woocommerce/base-hooks';
 
 /**
  * Internal dependencies
  */
-import { useShallowEqual } from './use-shallow-equal';
-import { usePrevious } from './use-previous';
+
+import { useQueryStateContext } from '../query-state-context';
 
 /**
  * A custom hook that exposes the current query state and a setter for the query
