+++ assets/js/base/hooks/index.js
@@ -2,16 +2,11 @@ export * from './checkout';
 export * from './customer';
 export * from './payment-methods';
 export * from './shipping';
-export * from './use-collection';
-export * from './use-collection-header';
-export * from './use-collection-data';
 export * from './use-container-queries';
 export * from './use-local-storage-state';
 export * from './use-position-relative-to-viewport';
 export * from './use-previous';
 export * from './use-shallow-equal';
-export * from './use-store-products';
 export * from './use-store-add-to-cart';
-export * from './use-query-state';
 export * from './use-throw-error';
 export * from './use-trigger-fragment-refresh';
